Refactor split smartdevice tests to test_iotdevice, test_smartdevice, and test_device
